requirements.txt
annotated-types==0.7.0
autocommand==2.2.2
backports.tarfile==1.2.0
bleach==6.2.0
blinker==1.9.0
blosc==1.11.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
flamegraph==0.1
Flask==3.1.0
fonttools==4.55.3
fsspec==2024.10.0
gevent==24.11.1
gprof2dot==2024.6.6
greenlet==3.1.1
h5py==3.12.1
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.4
joblib==1.4.2
kaggle==1.6.17
kiwisolver==1.4.7
linkify-it-py==2.0.3
llvmlite==0.43.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
memory-profiler==0.61.0
memray==1.15.0
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
psutil==6.1.1
py-spy==0.4.0
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pyinstrument==5.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scalene==1.5.49
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
snakeviz==2.2.2
sympy==1.13.1
tensorly==0.9.0
text-unidecode==1.3
textual==1.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
tomli==2.0.1
torch==2.5.1
-e git+https://github.com/pytorch/vision.git@06a925c32b49fd0455e265097fa7ca623cec4154#egg=torchvision
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.1.3
xxhash==3.5.0
zipp==3.19.2
zope.event==5.0
zope.interface==7.2
